+++ ReleaseNotes.md
@@ -6,7 +6,9 @@
 * 增加企业微信报警方式
 * Elasticsearch-Rest-Client升级至6.6.2
 * bugfix: 解决elasticsearch 7+版本，数据数量为0的问题
+* bugfix: 解决邮箱需要认证的情况下，邮件发送失败的问题
 * mysql: alert表增加字段: http_post_url
+* 文档: 增加query string简易教程
 
 ## 0.1-RELEASE